Re-check the storequeue after the action is run
This solves an issue where an action is waitingForState, but the state
change doesn't happen until the last action of a stack is fired. For
example:

1. say we had an action with something like `waitForState(state => state.loading === false);`
2. N actions fire that don't touch `state.loading`
3. one action, at the very end, sets `loading = false`

Because the storequeue was checked _before_ this last action, the
waitForState predicate never evaluates to true and so the action which
is waiting for `state.loading === false` is never dispatched.

I copypasted this code, which should be written into a function instead
of copypasted, to start discussion and confirm that this should be the
intended behavior.