Catch promise exceptions in `processQueue`.
Promise error handling is complex, this is a good example of such.
`processQueue` relies on `.then()` to continue process the queue
promises to resolve `PromseWell::onComplete`. If one of those
queue'd promises errors, the queue stalls and `node-platform`
based apps can hang on the server, leaving connections open
and users with a blank page forever (or until they hit 'stop').